+++ src/app/app.globals.ts
@@ -1,15 +1,15 @@
 // THIS IS AN AUTO-GENERATED FILE. DO NOT CHANGE IT MANUALLY!
-// Generated last time on Fri Aug 28 14:15:55 CEST 2020
+// Generated last time on Fri Oct  9 12:39:18 CEST 2020
 
 /**
  * The latest version of the AWG App
  */
-export const appVersion = '0.7.5';
+export const appVersion = '0.7.6';
 
 /**
  * The release date of the latest version of the AWG App
  */
-export const appVersionReleaseDate = '28. August 2020';
+export const appVersionReleaseDate = '09. Oktober 2020';
 
 /**
  * The URL of the AWG App