+++ src/app/shared/shared-ngbootstrap.module.ts
@@ -1,0 +1,35 @@
+import { NgModule } from '@angular/core';
+import {
+    NgbAccordionModule,
+    NgbButtonsModule,
+    NgbCollapseModule,
+    NgbDropdownModule,
+    NgbModalModule,
+    NgbNavModule,
+    NgbPaginationModule,
+    NgbTooltipModule
+} from '@ng-bootstrap/ng-bootstrap';
+
+@NgModule({
+    imports: [
+        NgbAccordionModule,
+        NgbButtonsModule,
+        NgbCollapseModule,
+        NgbDropdownModule,
+        NgbModalModule,
+        NgbNavModule,
+        NgbPaginationModule,
+        NgbTooltipModule
+    ],
+    exports: [
+        NgbAccordionModule,
+        NgbButtonsModule,
+        NgbCollapseModule,
+        NgbDropdownModule,
+        NgbModalModule,
+        NgbNavModule,
+        NgbPaginationModule,
+        NgbTooltipModule
+    ]
+})
+export class SharedNgbootstrapModule {}