+++ src/app/views/edition-view/edition-outlets/edition-graph/graph-visualizer/force-graph/d3/d3.service.ts
@@ -69,13 +69,7 @@ export class D3Service {
             d.fy = null;
         };
 
-        dragElement.call(
-            d3_drag
-                .drag()
-                .on('start', dragStart)
-                .on('drag', dragActions)
-                .on('end', dragEnd)
-        );
+        dragElement.call(d3_drag.drag().on('start', dragStart).on('drag', dragActions).on('end', dragEnd));
     }
 
     /**